+++ demo/src/client/faro/initialize.ts
@@ -15,12 +15,10 @@ export function initializeFaro(): Faro {
   const faro = coreInit({
     url: `http://localhost:${env.faro.portAppReceiver}/collect`,
     apiKey: env.faro.apiKey,
-
     instrumentations: [
       ...getWebInstrumentations({
         captureConsole: true,
       }),
-
       new TracingInstrumentation(),
       new ReactIntegration({
         router: {
@@ -41,13 +39,10 @@ export function initializeFaro(): Faro {
       version: env.package.version,
       environment: env.mode.name,
     },
-
     trackResources: true,
-
     batching: {
       itemLimit: 100,
     },
-
     trackUserActionsPreview: true,
   });