+++ packages/web-tracing/src/getDefaultOTELInstrumentations.ts
@@ -1,11 +1,24 @@
+import type { Span } from '@opentelemetry/api';
 import { FetchInstrumentation } from '@opentelemetry/instrumentation-fetch';
+
+import { stringifyExternalJson, unknownString } from '@grafana/faro-core';
+import { performanceEntriesSubscription } from '@grafana/faro-web-sdk';
 
 import { FaroXhrInstrumentation } from './faroXhrInstrumentation';
 import {
   fetchCustomAttributeFunctionWithDefaults,
+  type FetchError,
   xhrCustomAttributeFunctionWithDefaults,
 } from './instrumentationUtils';
 import type { DefaultInstrumentationsOptions, InstrumentationOption } from './types';
+
+// Workaround because the environment always maps the node type instead of the browser type
+type WithURL = {
+  url: string;
+};
+
+const xhrInitiatorType = 'xmlhttprequest';
+const fetchInitiatorType = 'fetch';
 
 export function getDefaultOTELInstrumentations(options: DefaultInstrumentationsOptions = {}): InstrumentationOption[] {
   const { fetchInstrumentationOptions, xhrInstrumentationOptions, ...sharedOptions } = options;
@@ -26,7 +39,12 @@ function createFetchInstrumentationOptions(
     ...fetchInstrumentationOptions,
     // always keep this function
     applyCustomAttributesOnSpan: fetchCustomAttributeFunctionWithDefaults(
-      fetchInstrumentationOptions?.applyCustomAttributesOnSpan
+      (span: Span, request: Request | RequestInit, result: Response | FetchError) => {
+        const url = (result as WithURL).url || (request as WithURL).url || getUrlFromSpan(span);
+        mapHttpRequestToPerformanceEntry(span, url, fetchInitiatorType);
+
+        fetchInstrumentationOptions?.applyCustomAttributesOnSpan?.(span, request, result);
+      }
     ),
   };
 }
@@ -41,8 +59,45 @@ function createXhrInstrumentationOptions(
     // keep this here to overwrite the defaults above if provided by the users
     ...xhrInstrumentationOptions,
     // always keep this function
-    applyCustomAttributesOnSpan: xhrCustomAttributeFunctionWithDefaults(
-      xhrInstrumentationOptions?.applyCustomAttributesOnSpan
-    ),
+    applyCustomAttributesOnSpan: xhrCustomAttributeFunctionWithDefaults((span: Span, xhr: XMLHttpRequest) => {
+      const url = xhr.responseURL || getUrlFromSpan(span);
+      mapHttpRequestToPerformanceEntry(span, url, xhrInitiatorType);
+
+      xhrInstrumentationOptions?.applyCustomAttributesOnSpan?.(span, xhr);
+    }),
   };
+}
+
+/**
+ * Map the request to the performance entry
+ *
+ * @remarks
+ * When instantiating your own FetchInstrumentation or XhrInstrumentation, call this function in the applyCustomAttributesOnSpan to ensure that the performance entry is mapped to the span.
+ *
+ * @param span - The span to map the request to
+ */
+export function mapHttpRequestToPerformanceEntry(span: Span, url: string, requestType: string) {
+  performanceEntriesSubscription.first().subscribe((msg) => {
+    const { faroNavigationId, faroResourceId, initiatorType, name } = msg.entry;
+    const isHttpRequestEntry = [xhrInitiatorType, fetchInitiatorType].includes(initiatorType);
+
+    if (!isHttpRequestEntry || name !== url || requestType !== initiatorType) {
+      return;
+    }
+
+    span.setAttribute('faro.performance.navigation.id', faroNavigationId);
+    span.setAttribute('faro.performance.resource.id', faroResourceId);
+  });
+}
+
+// Fallback to get the url from the span if not available in the request data.
+// This is a workaround until we have a better solution.
+// Maybe we should implement our own FetchInstrumentation and XhrInstrumentation where can hook into the request and response data.
+function getUrlFromSpan(span: Span) {
+  try {
+    const parsedSpan = JSON.parse(stringifyExternalJson({ ...span }));
+    return parsedSpan.attributes['http.url'];
+  } catch (e) {
+    return unknownString;
+  }
 }