+++ packages/web-sdk/src/instrumentations/csp/index.ts
@@ -1,0 +1,1 @@
+export { CSPInstrumentation } from './instrumentation';