+++ CHANGELOG.md
@@ -7,8 +7,8 @@
 - chore(user actions): Rename `userActionEventType` to `userActionTrigger` for improved clarity (#1298)
 
 - chore(`@grafana/faro web-sdk`): Move the stackFrames from the errors instrumentation package to the utils package (#1316)
-- improvement (`@grafana/faro-web-sdk`): Allow users to configure the stack frame tracer in the ErrorsInstrumentation and the stackTraceParser in the BrowserConfig (#1316)
-
+- improvement (`@grafana/faro-web-sdk`): Allow users to configure the stack frame tracer
+  in the ErrorsInstrumentation and the stackTraceParser in the BrowserConfig (#1316)
 
 ## 1.18.2